racket/gui: fix reporting of numpad Enter as 'numpad-enter
On Unix and Mac OS X, the key has long been reported as #\u3,
even though the documentation says 'numpad-enter.

A more compatible change would be to adjust the documentation and
also change DrRacket and the framework to recognize #\u3 instead
of 'numpad-enter. But there doesn't seem to be a good reason to
treat #\u3 as "numpad Enter" --- particularly as opposed to treating
it as "Ctl-C" --- other than ancient Mac history.

On Windows, numpad Enter continues to be reported as #\return,
since Windows event handling doesn't naturally or easily distinguish
the two keys.