DrRacket: drop numpad Enter treatment as force-evaluation
DrRacket's interactions window tried to treat numpad Enter the same
as ctl-Return: send input to the reader, even if an S-expression is
incomplete. Since numpad Enter events were not reported as
'numpad-enter, however, the special treatment didn't take effect.

Meanwhile, it appears that some keyboard configurations report a
numpad Enter event when NumLock is on and the regular Enter key
is pressed, and that's probably not what a user wants. And
numpad Enter has never been reported differently from regular
Enter on Windows.

So, it's simpler and more consistent to just treat numpad Enter
and regular Enter the same.