fix to unit macro
Here's example that's worrying --- works in the current Racket, but
not here:

     #lang racket

     (define-syntax-rule (define1 id)
      (begin
        (define x 5)
        (define-syntax id #'x)))

     (define-syntax (use stx)
       (syntax-case stx ()
         [(_ id)
          (with-syntax ([old-id (syntax-local-value #'id)])
            #'(begin
                (define x 6)
                ;; reference to old-id ends up ambigious
                (list x old-id)))]))

     (define1 foo)
     (use foo)

The problem is that the `x` reference is introduced by two macro
applications, while each of those introduces an `x`. It works in
Racket currently because mark order matters.

Then again, this example doesn't work in Racket currently:

     #lang racket

     (begin
       (define x 5)
       (define-syntax foo #'x))

     (define-syntax (use stx)
       (syntax-case stx ()
         [(_ id)
          (with-syntax ([old-id (syntax-local-value #'id)])
            #'(begin
                (define x 6)
                (list x old-id)))]))

     (use foo)

Given this example, I'm inclined to say that the `use` macro is at
fault, and it's nice in a sense that the new macro system exposes the
problem earlier.

The general rule is that whenever a macro pulls in an id for
reference, it should unintroduce it. It looks like the unit
macors mostly do that, but not for an identifier that holds
a unit value.