restore expander's substitution of binding ids for references
Being able to use `bound-identifier=?` is not only more compatible
with the old expander, it is more convenient for transformations that
interpose capturing bindings on an expression based on it's free
variables. (If there is more than one reference to a given binding,
the references might have some scopes not in common, so neither would
bind the other). Unlike the old expander, though, this substitution
must not happen when `local-expand` is used with a stop list.

Given the change to the expander, we switch `syntax/free-vars` back
to using `bound-identifier=?`.