preserve full binding info
Preserve enoug binding information so that, for example,

 (define x (make-rename-transformer #'y))

can install a `free-id=?` mapping when it is run, so that re-running
a module can reset any shadowing put in place via `module->namespace`,
and so on.

This change keeps binding information in more modules than we'd
like, but let's get everything working and then look for
representation improvements and shortcuts.