checkpoint: new approach to top-level contexts
Streamline module-context marks and introduce a new approach to
the top-level context, which aims to allow syntax objects
and compiled code from one top-level context to be used
(in essentially the old way) in a different top-level context.

It's not quite right (a "stx.rktl" test fails), but it's getting close.