change scope distinction between macro defn and use
In a definition context, mark syntax going into a macro so that both
original and introduced syntax have distingiushing marks. This approach
replaces the idea of trying to mark everything in an module or definition
context preemptively with an "expression" mark; that doesn't work, because
a similar distinction is needed when a macro expansion introduces
both definitions and uses of a new macro.