make binding addition detect simplication opportunity via shared table
When bindings exported by a module are individually added,
detect when all of the bindings have been added and convert to
the shorthand form tat uses a shared export table from the module.

This simplification is particularly useful for TR's implementation,
which uses a require transformer that expands to individual imports
(because that's the require transformer API).

Also, track `module->namespace` context for a `(module* .... #f ....)`
submodule as the context of the #f.