more compact representation of HAMTs
For an `eq?`-based table, it's wasteful to store the hash code in the
table, since it can be recomputed from the key. And if we really just
want a set, then it's wasteful to store a value in addition to a key.
Reorder the content so that the value and code can be optional.

Change mark sets to use `eq?`-based set HAMTs.