remove `syntax-local-make-delta-introducer`; change `unit` binding
Also, eliminate the use of `syntax-local-get-shaodwer` except
in `syntax-parameterize` (and the lightweight variant in `class`).

The binding of signature elements is now determined by the
context of the siganture reference, instead of using
`syntax-local-get-shadower`.

The reason to purge `syntax-local-get-shadower` is that it's
difficult to define across module boundaries. The old implementation
stripped "module context" from its argument, but that concept
is not well defined, particularly when contexts from multiple
modules or top levels are involved  through `eval`.

Identifiers within a signature can have extra marks relative to
the binding of the signature, in which case the extra marks
are preserved when the signature's elements are unpacked to
new bindings. That perservation of mark delts preserved the
sense in which signature elements are identifiers, not symbols.