simplify `syntax-local-get-shadower`
Make the oeration less powerful --- just powerful enough to
implement syntax parameters.

An alternative would be to build something even closer to syntax
parameters into the expander, and get rid of
`syntax-local-get-shadower`. After experimenting with some interfaces,
however, a simplified `syntax-local-get-shadower` seems to be as clear
and convenient as anything.