repair `define/generic`
The test that I "corrected" before was ok. My ideas about binding
weren't consistent, and the old behavior seems the most consistent
choice after all.

Method names in a generic are semi-symbolic: the binding of the
name doesn't matter, but any extra scopes that are on the name
do matter. That means that methods cannot be renamed on export
and have the rename affect uses of `#:methods` or `define/generic`,
but it also means that macro can introduce methods that are
accessible only by macro-introduced references.